Core Operations

Core Operations

Our ability to cast the input-output constraints of a neural net to an efficient set of linear and integer constraints boils down to the following basic operations, over which the layers provide a convenient layer of abstraction.

Index

MIPVerify.abs_ge

MIPVerify.masked_relu

MIPVerify.maximum

MIPVerify.relu

MIPVerify.set_max_indexes

Internal

MIPVerify.abs_ge — Method.

abs_ge(x)

Expresses a one-sided absolute-value constraint: output is constrained to be at least as large as |x|.

Only use when you are minimizing over the output in the objective.

source

MIPVerify.masked_relu — Method.

masked_relu(x, m)

Expresses a masked rectified-linearity constraint, with three possibilities depending on the value of the mask. Output is constrained to be:

1) max(x, 0) if m=0, 
2) 0 if m<0
3) x if m>0

source

MIPVerify.maximum — Method.

maximum(xs; tighten)

Expresses a maximization constraint: output is constrained to be equal to max(xs).

source

MIPVerify.relu — Method.

relu(x)

Expresses a rectified-linearity constraint: output is constrained to be equal to max(x, 0).

source

MIPVerify.set_max_indexes — Method.

set_max_indexes(x, target_indexes; tolerance)

Imposes constraints ensuring that one of the elements at the target_indexes is the largest element of the array x. More specifically, we require x[j] - x[i] ≥ tolerance for some j ∈ target_indexes and for all i ∉ target_indexes.

source