Layers

Layers

Each layer in the neural net corresponds to a struct that simultaneously specifies: 1) the operation being carried out in the layer (recorded in the _type_ of the struct) and 2) the parameters for the operation (recorded in the values of the fields of the struct).

When we pass an input array of real numbers to a layer struct, we get an output array of real numbers that is the result of the layer operating on the input.

Conversely, when we pass an input array of JuMP variables, we get an output array of JuMP variables, with the appropriate mixed-integer constraints (as determined by the layer) imposed between the input and output.

Index

MIPVerify.MaxPoolParameters

MIPVerify.conv2d

MIPVerify.convolution_layer

MIPVerify.fully_connected_layer

MIPVerify.getoutputsize

MIPVerify.getpoolview

MIPVerify.getsliceindex

MIPVerify.masked_fully_connected_layer

MIPVerify.matmul

MIPVerify.pool

MIPVerify.poolmap

MIPVerify.Conv2DParameters

MIPVerify.Conv2DParameters

MIPVerify.ConvolutionLayerParameters

MIPVerify.ConvolutionLayerParameters

MIPVerify.FullyConnectedLayerParameters

MIPVerify.MaskedFullyConnectedLayerParameters

MIPVerify.MatrixMultiplicationParameters

MIPVerify.PoolParameters

MIPVerify.SoftmaxParameters

Public Interface

MIPVerify.Conv2DParameters — Type.

struct Conv2DParameters{T<:Union{JuMP.AbstractJuMPScalar, Real}, U<:Union{JuMP.AbstractJuMPScalar, Real}} <: MIPVerify.LayerParameters

Stores parameters for a 2-D convolution operation.

p(x) is shorthand for conv2d(x, p) when p is an instance of Conv2DParameters.

Fields:

filter

bias

source

MIPVerify.Conv2DParameters — Method.

Conv2DParameters(filter)

Convenience function to create a Conv2DParameters struct with the specified filter and zero bias.

source

MIPVerify.ConvolutionLayerParameters — Type.

struct ConvolutionLayerParameters{T<:Real, U<:Real} <: MIPVerify.StackableLayerParameters

Stores parameters for a convolution layer consisting of a convolution, followed by max-pooling, and a ReLU activation function.

p(x) is shorthand for convolution_layer(x, p) when p is an instance of ConvolutionLayerParameters.

Fields:

conv2dparams

maxpoolparams

source

MIPVerify.ConvolutionLayerParameters — Method.

ConvolutionLayerParameters(filter, bias, strides)

Convenience function to create a ConvolutionLayerParameters struct with the specified filter, bias, and strides for the max-pooling operation.

source

MIPVerify.FullyConnectedLayerParameters — Type.

struct FullyConnectedLayerParameters{T<:Real, U<:Real} <: MIPVerify.StackableLayerParameters

Stores parameters for a fully connected layer consisting of a matrix multiplication followed by a ReLU activation function.

p(x) is shorthand for fully_connected_layer(x, p) when p is an instance of FullyConnectedLayerParameters.

Fields:

mmparams

source

MIPVerify.MaskedFullyConnectedLayerParameters — Type.

struct MaskedFullyConnectedLayerParameters{T<:Real, U<:Real, V<:Real} <: MIPVerify.StackableLayerParameters

Stores parameters for a fully connected layer consisting of a matrix multiplication followed by a ReLU activation function that is activated selectively depending on the corresponding value of the mask.

p(x) is shorthand for masked_fully_connected_layer(x, p) when p is an instance of MaskedFullyConnectedLayerParameters.

source

MIPVerify.MatrixMultiplicationParameters — Type.

struct MatrixMultiplicationParameters{T<:Real, U<:Real} <: MIPVerify.LayerParameters

Stores parameters for a layer that does a simple matrix multiplication.

p(x) is shorthand for matmul(x, p) when p is an instance of MatrixMultiplicationParameters.

Fields:

matrix

bias

source

MIPVerify.MaxPoolParameters — Method.

MaxPoolParameters(strides)

Convenience function to create a PoolParameters struct for max-pooling.

source

MIPVerify.PoolParameters — Type.

struct PoolParameters{N} <: MIPVerify.LayerParameters

Stores parameters for a pooling operation.

p(x) is shorthand for pool(x, p) when p is an instance of PoolParameters.

Fields:

strides

pooling_function

source

MIPVerify.SoftmaxParameters — Type.

struct SoftmaxParameters{T<:Real, U<:Real} <: MIPVerify.LayerParameters

Stores parameters for a softmax layer consisting of a matrix multiplication with _no_ activation function.

This simply wraps MatrixMultiplicationParameters to ensure that it is distinguishable as a softmax layer.

Fields:

mmparams

source

Internal

MIPVerify.conv2d — Method.

conv2d(input, params)

Computes the result of convolving input with the filter and bias stored in params.

Mirrors tf.nn.conv2d from the tensorflow package, with strides = [1, 1, 1, 1], padding = 'SAME'.

Throws

AssertionError if input and filter are not compatible.

source

MIPVerify.convolution_layer — Method.

convolution_layer(x, params)

Computes the result of convolving x with params.conv2dparams, pooling the resulting output with the pooling function and strides specified in params.maxpoolparams, and passing the output through a ReLU activation function.

source

MIPVerify.fully_connected_layer — Method.

Computes the result of multiplying x by params.mmparams, and passing the output through a ReLU activation function.

source

MIPVerify.masked_fully_connected_layer — Method.

Similar to fully_connected_layer(x, params), but with an additional mask in params that controls whether a ReLU is applied to each output.

If the value of the mask is <0 (i.e. input is assumed to be always non-positive), the output is set at 0.

If the value of the mask is 0 (i.e. input can take both positive and negative values), the output is rectified.

If the value of the mask is >0 (i.e. input is assumed to be always non-negative), the output is set as the value of the input, without any rectification.

source

MIPVerify.matmul — Method.

matmul(x, params)

Computes the result of pre-multiplying x by the transpose of params.matrix and adding params.bias.

source

MIPVerify.getoutputsize — Method.

getoutputsize(input_array, strides)

For pooling operations on an array, returns the expected size of the output array.

source

MIPVerify.getpoolview — Method.

getpoolview(input_array, strides, output_index)

For pooling operations on an array, returns a view of the parent array corresponding to the output_index in the output array.

source

MIPVerify.getsliceindex — Method.

getsliceindex(input_array_size, stride, output_index)

For pooling operations on an array where a given element in the output array corresponds to equal-sized blocks in the input array, returns (for a given dimension) the index range in the input array corresponding to a particular index output_index in the output array.

Returns an empty array if the output_index does not correspond to any input indices.

Arguments

stride::Integer: the size of the operating blocks along the active dimension.

source

MIPVerify.pool — Method.

pool(input, params)

Computes the result of applying the pooling function params.pooling_function to non-overlapping cells of input with sizes specified in params.strides.

source

MIPVerify.poolmap — Method.

poolmap(f, input_array, strides)

Returns output from applying f to subarrays of input_array, with the windows determined by the strides.

source