Networks

Networks

Each network corresponds to an array of layers associated with a unique string identifier. The string identifier of the network is used to store cached models, so it's important to ensure that you don't re-use names!

Public Interface

MIPVerify.MaskedFullyConnectedNetParameters — Type.

struct MaskedFullyConnectedNetParameters <: MIPVerify.NeuralNetParameters

Represents a neural net consisting of multiple masked fully-connected layers (as an array of MaskedFullyConnectedLayerParameters), followed by a softmax layer (as a SoftmaxParameters).

Fields:

masked_fclayer_params

softmax_params

UUID

source

MIPVerify.StandardNeuralNetParameters — Type.

struct StandardNeuralNetParameters <: MIPVerify.NeuralNetParameters

Represents a neural net consisting of multiple convolution layers (as an array of ConvolutionLayerParameters), followed by multiple fully-connected layers (as an array of FullyConnectedLayerParameters), followed by a softmax layer (as a SoftmaxParameters).

You can leave the array convlayer_params empty if you do not have convolution layers, or conversely leave fclayers_empty if you do not have fully-connected layers. (Leaving _both_ empty doesn't make sense!)

Fields:

convlayer_params

fclayer_params

softmax_params

UUID

source

Internal