Overview

Overview

A neural net consists of multiple layers, each of which (potentially) operates on input differently. We represent these objects with NeuralNetParameters and LayerParameters.

Index

MIPVerify.LayerParameters

MIPVerify.NeuralNetParameters

MIPVerify.StackableLayerParameters

PublicInterface

MIPVerify.LayerParameters — Type.

abstract LayerParameters

Supertype for all types storing the parameters of each layer. Inherit from this to specify your own custom type of layer. Each implementation is expected to:

Implement a callable specifying the output when any input of type JuMPReal is provided.

source

MIPVerify.StackableLayerParameters — Type.

abstract StackableLayerParameters <: MIPVerify.LayerParameters

Supertype for all LayerParameters that can be logically applied in sequence.

An array of StackableLayerParameters is interpreted as that array of layer being applied to the input sequentially, starting from the leftmost layer. (In functional programming terms, this can be thought of as a sort of fold).

source

MIPVerify.NeuralNetParameters — Type.

abstract NeuralNetParameters

Supertype for all types storing the parameters of a neural net. Inherit from this to specify your own custom architecture of LayerParameters. Each implementation is expected to:

Implement a callable specifying the output when any input of type JuMPReal is provided

Have a UUID field for the name of the neural network.

source

MIPVerify.JuMPReal — Constant.

source