Importing External Datasets

Importing External Datasets

For your convenience, the MNIST dataset is available as part of our package.

Index

MIPVerify.read_datasets

MIPVerify.LabelledImageDataset

MIPVerify.NamedTrainTestDataset

Public Interface

MIPVerify.read_datasets — Method.

read_datasets(name)

Makes popular machine learning datasets available as a NamedTrainTestDataset.

Arguments

name::String: name of machine learning dataset. Options:

MNIST: The MNIST Database of handwritten digits

source

Internal

MIPVerify.LabelledImageDataset — Type.

struct LabelledImageDataset{T<:Real, U<:Int64} <: MIPVerify.LabelledDataset

Dataset of images stored as a 4-dimensional array of size (num_samples, image_height, image_width, num_channels), with accompanying labels (sorted in the same order) of size num_samples.

source

MIPVerify.NamedTrainTestDataset — Type.

struct NamedTrainTestDataset{T<:MIPVerify.Dataset} <: MIPVerify.Dataset

Named dataset containing a training set and a test set which are expected to contain the same kind of data.

source