Batch Processing

Batch Processing

When running on multiple samples from a single dataset, batch_find_certificate allows you to redo solves intelligently - redoing 1) no solves, 2) all solves, 3) only solves where the sample status is indeterminate, or 4) only solves where the best counter-example is non-optimal.

Index

MIPVerify.batch_find_certificate

MIPVerify.run_on_sample

Public Interface

MIPVerify.batch_find_certificate — Method.

batch_find_certificate(nn, dataset, target_sample_numbers, main_solver; save_path, pp, norm_order, tolerance, rebuild, tightening_algorithm, tightening_solver, solve_rerun_option, cache_model)

Runs find_adversarial_example for the specified neural network nn and dataset for the target_sample_numbers, skipping target_sample_numbers based on the selected solve_rerun_option.

It creates a named directory in save_path, with the name summarizing

the name of the network in nn,

the perturbation family pp,

the norm_order

the tolerance.

Within this directory, a summary of all the results is stored in summary.csv, and results from individual runs are stored in the subfolder run_results.

This functioned is designed so that it can be interrupted and restarted cleanly; it relies on the summary.csv file to determine what the results of previous runs are (so modifying this file manually can lead to unexpected behavior.)

main_solver specifies the solver used.

Named Arguments:

pp, norm_order, tolerance, rebuild, tightening_algorithm, tightening_solver are passed directly to find_adversarial_example; see that documentation for more details.

pp::PerturbationFamily: Defaults to UnrestrictedPerturbationFamily().

norm_order::Real: Defaults to 1.

tolerance::Real: Defaults to 0.0.

rebuild::Bool: Defaults to false.

tightening_algorithm::MIPVerify.TighteningAlgorithm: Defaults to mip.

tightening_solver:

solve_rerun_option::MIPVerify.SolveRerunOption: Options are never, always, resolve_ambiguous_cases, and refine_insecure_cases. See run_on_sample for more details.

source

Internal

MIPVerify.run_on_sample — Method.

run_on_sample(sample_number, summary_dt, solve_rerun_option)

Determines whether to run a solve on a sample depending on the solve_rerun_option by looking up information on the most recent completed solve. recorded in summary_dt

summary_dt is expected to be a DataFrame with columns :SampleNumber, :SolveStatus, and :ObjectiveValue.

Behavior for different choices of solve_rerun_option:

never: true if and only if there is no previous completed solve.

always: true always.

resolve_ambiguous_cases: true if there is no previous completed solve, or if the most recent completed solve a) did not find a counter-example BUT b) the optimization was not demosntrated to be infeasible.

refine_insecure_cases: true if there is no previous completed solve, or if the most recent complete solve a) did find a counter-example BUT b) we did not reach a provably optimal solution.

source