Core Operations

Core Operations

Our ability to cast the input-output constraints of a neural net to an efficient set of linear and integer constraints boils down to the following basic operations, over which the layers provide a convenient layer of abstraction.

Index

MIPVerify.abs_ge

MIPVerify.lazy_tight_lowerbound

MIPVerify.masked_relu

MIPVerify.maximum

MIPVerify.maximum_ge

MIPVerify.relu

MIPVerify.set_max_indexes

MIPVerify.tight_bound

Internal

MIPVerify.abs_ge — Method.

abs_ge(x)

Expresses a one-sided absolute-value constraint: output is constrained to be at least as large as |x|.

Only use when you are minimizing over the output in the objective.

source

MIPVerify.lazy_tight_lowerbound — Method.

Calculates the lowerbound only if u is positive; otherwise, returns u (since we expect) the ReLU to be fixed to zero anyway.

source

MIPVerify.masked_relu — Method.

masked_relu(x, m; nta)

Expresses a masked rectified-linearity constraint, with three possibilities depending on the value of the mask. Output is constrained to be:

1) max(x, 0) if m=0, 
2) 0 if m<0
3) x if m>0

source

MIPVerify.maximum — Method.

maximum(xs)

Expresses a maximization constraint: output is constrained to be equal to max(xs).

source

MIPVerify.maximum_ge — Method.

maximum_ge(xs)

Expresses a one-sided maximization constraint: output is constrained to be at least max(xs).

Only use when you are minimizing over the output in the objective.

source

MIPVerify.relu — Method.

relu(x)
relu(x; nta)

Expresses a rectified-linearity constraint: output is constrained to be equal to max(x, 0).

source

MIPVerify.set_max_indexes — Method.

set_max_indexes(x, target_indexes; tolerance)

Imposes constraints ensuring that one of the elements at the target_indexes is the largest element of the array x. More specifically, we require x[j] - x[i] ≥ tolerance for some j ∈ target_indexes and for all i ∉ target_indexes.

source

MIPVerify.tight_bound — Method.

Calculates a tight bound of type bound_type on the variable x using the specified tightening algorithm nta.

If an upper bound is proven to be below cutoff, or a lower bound is proven to above cutoff, the algorithm returns early with whatever value was found.

source