Layers

Layers

Each layer in the neural net corresponds to a struct that simultaneously specifies: 1) the operation being carried out in the layer (recorded in the type of the struct) and 2) the parameters for the operation (recorded in the values of the fields of the struct).

When we pass an input array of real numbers to a layer struct, we get an output array of real numbers that is the result of the layer operating on the input.

Conversely, when we pass an input array of JuMP variables, we get an output array of JuMP variables, with the appropriate mixed-integer constraints (as determined by the layer) imposed between the input and output.

Index

MIPVerify.MaxPool

MIPVerify.conv2d

MIPVerify.flatten

MIPVerify.getoutputsize

MIPVerify.getpoolview

MIPVerify.getsliceindex

MIPVerify.matmul

MIPVerify.matmul

MIPVerify.pool

MIPVerify.poolmap

MIPVerify.Conv2d

MIPVerify.Pool

MIPVerify.ReLU

Public Interface

MIPVerify.Conv2d — Method.

Conv2d(filter)

Convenience function to create a Conv2d struct with the specified filter and zero bias.

MIPVerify.MaxPool — Method.

MaxPool(strides)

Convenience function to create a Pool struct for max-pooling.

MIPVerify.Pool — Type.

struct Pool{N} <: Layer

Represents a pooling operation.

p(x) is shorthand for pool(x, p) when p is an instance of Pool.

Fields:

strides

pooling_function

MIPVerify.ReLU — Type.

struct ReLU <: Layer

Represents a ReLU operation.

p(x) is shorthand for relu(x) when p is an instance of ReLU.

Internal

MIPVerify.conv2d — Method.

conv2d(input, params)

Computes the result of convolving input with the filter and bias stored in params.

Mirrors tf.nn.conv2d from the tensorflow package, with strides = [1, 1, 1, 1], padding = 'SAME'.

Throws

AssertionError if input and filter are not compatible.

MIPVerify.flatten — Method.

Permute dimensions of array in specified order, then flattens the array.

MIPVerify.matmul — Method.

matmul(x, params)

Computes the result of pre-multiplying x by the transpose of params.matrix and adding params.bias.

MIPVerify.matmul — Method.

matmul(x, params)

Computes the result of pre-multiplying x by the transpose of params.matrix and adding params.bias. We write the computation out by hand when working with JuMPLinearType so that we are able to simplify the output as the computation is carried out.

MIPVerify.getoutputsize — Method.

getoutputsize(input_array, strides)

For pooling operations on an array, returns the expected size of the output array.

MIPVerify.getpoolview — Method.

getpoolview(input_array, strides, output_index)

For pooling operations on an array, returns a view of the parent array corresponding to the output_index in the output array.

MIPVerify.getsliceindex — Method.

getsliceindex(input_array_size, stride, output_index)

For pooling operations on an array where a given element in the output array corresponds to equal-sized blocks in the input array, returns (for a given dimension) the index range in the input array corresponding to a particular index output_index in the output array.

Returns an empty array if the output_index does not correspond to any input indices.

Arguments

stride::Integer: the size of the operating blocks along the active dimension.

MIPVerify.pool — Method.

pool(input, params)

Computes the result of applying the pooling function params.pooling_function to non-overlapping cells of input with sizes specified in params.strides.

MIPVerify.poolmap — Method.

poolmap(f, input_array, strides)

Returns output from applying f to subarrays of input_array, with the windows determined by the strides.