Networks

Networks

Each network corresponds to an array of layers associated with a unique string identifier. The string identifier of the network is used to store cached models, so it's important to ensure that you don't re-use names!

Public Interface

MIPVerify.Sequential — Type.

struct Sequential <: MIPVerify.NeuralNet

Represents a sequential (feed-forward) neural net, with layers ordered from input to output.

Fields:

layers

UUID

source

Internal