Overview

A neural net consists of multiple layers, each of which (potentially) operates on input differently. We represent these objects with NeuralNet and Layer.

Index

MIPVerify.chain

MIPVerify.Layer

MIPVerify.NeuralNet

PublicInterface

MIPVerify.chain — Method

An array of Layers is interpreted as that array of layer being applied to the input sequentially, starting from the leftmost layer. (In functional programming terms, this can be thought of as a sort of fold).

MIPVerify.Layer — Type

abstract type Layer

Supertype for all types storing the parameters of each layer. Inherit from this to specify your own custom type of layer. Each implementation is expected to:

Implement a callable specifying the output when any input of type JuMPReal is provided.

MIPVerify.NeuralNet — Type

abstract type NeuralNet

Supertype for all types storing the parameters of a neural net. Inherit from this to specify your own custom architecture. Each implementation is expected to:

Implement a callable specifying the output when any input of type JuMPReal is provided

Have a UUID field for the name of the neural network.