Datasets

For your convenience, the MNIST and CIFAR10 dataset is available as part of our package.

Index

MIPVerify.read_datasets

MIPVerify.LabelledImageDataset

MIPVerify.NamedTrainTestDataset

Public Interface

MIPVerify.read_datasets — Method

read_datasets(name)

Makes popular machine learning datasets available as a NamedTrainTestDataset.

Arguments

name::String: name of machine learning dataset. Options:

MNIST: The MNIST Database of handwritten digits. Pixel values in original dataset are provided as uint8 (0 to 255), but are scaled to range from 0 to 1 here.

CIFAR10: Labelled subset in 10 classes of 80 million tiny images dataset. Pixel values in original dataset are provided as uint8 (0 to 255), but are scaled to range from 0 to 1 here.

Internal

MIPVerify.LabelledImageDataset — Type

struct LabelledImageDataset{T<:Real, U<:Integer} <: MIPVerify.LabelledDataset

Dataset of images stored as a 4-dimensional array of size (num_samples, image_height, image_width, num_channels), with accompanying labels (sorted in the same order) of size num_samples.

MIPVerify.NamedTrainTestDataset — Type

struct NamedTrainTestDataset{T<:MIPVerify.Dataset, U<:MIPVerify.Dataset} <: MIPVerify.Dataset

Named dataset containing a training set and a test set which are expected to contain the same kind of data.

name: Name of dataset.

train: Training set.

test: Test set.